protectedRoute has been removed, routes for the app have been changed, routes have been removed, fixed lint and prettier errors
