+++ src/router/authorization.tsx
@@ -1,9 +1,9 @@
 import type { FC } from 'react'
-import { Navigate } from 'react-router-dom'
+import { Navigate, Outlet } from 'react-router-dom'
 import { observer } from 'mobx-react-lite'
 import authStore from '@/stores/authStore'
 import type { AuthorizationProps } from './types'
-import { ROUTES } from './routes'
+import { ROUTES } from '@/router/constants.ts'
 
 export const Authorization: FC<AuthorizationProps> = observer(
   ({ requireAuth }) => {
@@ -17,5 +17,7 @@ export const Authorization: FC<AuthorizationProps> = observer(
     if (requireAuth && !isAuth) {
       return <Navigate to={ROUTES.signIn} replace />
     }
+
+    return <Outlet />
   }
 )
