+++ src/router/constants.ts
@@ -1,0 +1,8 @@
+export const ROUTES = {
+  home: '/',
+  admin: (page = '') => `/admin/${page}`,
+  app: (page = '') => `/app/${page}`,
+  signIn: '/sign-in',
+  signUp: '/sign-up',
+  signOut: '/sign-out',
+}
