+++ src/router/index.tsx
@@ -1,8 +1,6 @@
 import { createBrowserRouter } from 'react-router-dom'
-import {routes} from './routes'
+import { routes } from './routes'
 
-const router = createBrowserRouter([
-  ...routes
-])
+const router = createBrowserRouter([...routes])
 
 export default router
